requirements.txt
lxml>=4.9.0
click>=8.1.0
adb-shell>=0.4.0
pytest>=7.0.0
python-dotenv>=1.0.0
requests>=2.31.0
# Optional: For Vertex AI (more complex setup)
# google-cloud-aiplatform>=1.38.0